pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "backtesting-engine"
version = "0.1.0"
description = "Simple backtesting and risk evaluation engine"
authors = [{ name = "Andrew Geday" }]
dependencies = [
  "pandas",
  "numpy",
  "pyyaml"
]